requirements.txt
flask
pillow
torch
torchvision
gdown
flask-cors
